+++ src/widgets/list/node.rs
@@ -96,8 +96,11 @@ impl WidgetNode for ListNode {
         !self.disabled
             && (self.on_click.is_some()
                 || self.on_select.is_some()
+                || self.on_item_click.is_some()
+                || self.on_activate.is_some()
                 || self.on_scroll_to.is_some()
                 || self.scrollbar
+                || self.show_scroll_indicators
                 || self.scroll_wheel)
     }